Correct bugs with merging
